feat(bsppp): add BSP editing library, use as the backend of the BSP format in vpkpp
