+++ src/bsppp/_bsppp.cmake
@@ -1,0 +1,3 @@
+add_pretty_parser(bsppp DEPS vpkpp SOURCES
+        "${CMAKE_CURRENT_SOURCE_DIR}/include/bsppp/bsppp.h"
+        "${CMAKE_CURRENT_LIST_DIR}/bsppp.cpp")
